+++ app/proguard-rules.pro
@@ -1,2 +1,2 @@
--keep class info.simplyapps.appengine.*
--keep class info.simplyapps.gameengine.*
+-keep class com.juergenkleck.android.appengine.*
+-keep class com.juergenkleck.android.gameengine.*
