+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="info.simplyapps.game.feedthecat.free"
+    package="com.juergenkleck.android.game.feedthecat.free"
     android:installLocation="auto" >
 
     <supports-screens android:resizeable="true"
@@ -19,11 +19,12 @@
         android:theme="@style/AppTheme">
 
         <activity
-            android:name="info.simplyapps.game.feedthecat.screens.HomeScreen"
+            android:name="com.juergenkleck.android.game.feedthecat.screens.HomeScreen"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar"
             android:configChanges="orientation|keyboardHidden"
-            android:screenOrientation="landscape" >
+            android:screenOrientation="landscape"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
