+++ app/src/main/java/com/juergenkleck/android/game/feedthecat/engine/FeedTheCatEngine.java
@@ -1,0 +1,12 @@
+package com.juergenkleck.android.game.feedthecat.engine;
+
+import com.juergenkleck.android.gameengine.engine.GameEngine;
+
+/**
+ * Android app - FeedTheCat
+ *
+ * Copyright 2022 by Juergen Kleck <develop@juergenkleck.com>
+ */
+public interface FeedTheCatEngine extends GameEngine {
+
+}
