+++ app/src/main/java/com/juergenkleck/android/game/feedthecat/rendering/objects/Obstacle.java
@@ -1,0 +1,14 @@
+package com.juergenkleck.android.game.feedthecat.rendering.objects;
+
+/**
+ * Android app - FeedTheCat
+ *
+ * Copyright 2022 by Juergen Kleck <develop@juergenkleck.com>
+ */
+public class Obstacle extends com.juergenkleck.android.gameengine.rendering.objects.Obstacle {
+
+    public Obstacle(int type, float life, int gReference, int points) {
+        super(type, life, gReference, points);
+    }
+
+}
