--- app/src/main/java/info/simplyapps/game/feedthecat/engine/FeedTheCatEngine.java
@@ -1,7 +1,0 @@
-package info.simplyapps.game.feedthecat.engine;
-
-import info.simplyapps.gameengine.engine.GameEngine;
-
-public interface FeedTheCatEngine extends GameEngine {
-
-}
