--- app/src/main/java/info/simplyapps/game/feedthecat/rendering/objects/Obstacle.java
@@ -1,9 +1,0 @@
-package info.simplyapps.game.feedthecat.rendering.objects;
-
-public class Obstacle extends info.simplyapps.gameengine.rendering.objects.Obstacle {
-
-    public Obstacle(int type, float life, int gReference, int points) {
-        super(type, life, gReference, points);
-    }
-
-}
